+++ lib/src/hosts/azure_host.dart
@@ -54,6 +54,7 @@ class AzureHost extends HostAdapter {
   List<DatabaseProvider> get supportedDatabases => const [
         DatabaseProvider.none,
         DatabaseProvider.neon,
+        DatabaseProvider.supabase,
       ];
 
   @override